requirements.txt
aiogram
aiosqlite
python-decouple